requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-storage-blob==12.25.0
backports.zoneinfo==0.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
diff-match-patch==20241021
Django==4.2.20
django-environ==0.11.2
django-import-export==4.1.1
django-mssql-backend==2.8.1
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==2.0.0
idna==3.10
isodate==0.7.2
mssql-django==1.5
openpyxl==3.1.5
pycparser==2.22
PyJWT==2.9.0
pyodbc==5.2.0
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
urllib3==2.2.3
